requirements.txt
streamlit
ultralytics
torch
numpy
pandas
scikit-learn
joblib
plotly
pillow
prophet
python-dotenv
